# openapi.yaml
openapi: 3.0.3
info:
  title: Sectors News API
  description: |
    A comprehensive API for Indonesian stock market news processing, analysis, and management.

    ## Features
    - **News Article Processing**: Extract, classify, and score news articles from URLs
    - **Insider Trading Filings**: Process PDF filings and insider trading data
    - **Performance Monitoring**: Real-time performance metrics and monitoring
    - **Subscription Management**: User subscription management for news topics
    - **Content Classification**: Automatic tagging, ticker extraction, and sentiment analysis

    ## Authentication
    All endpoints (except GET /articles) require an API key passed in the `X-API-Key` header.

    ## Performance Optimizations
    The API includes advanced performance optimizations including:
    - Intelligent caching for URL content
    - Parallel processing for LLM operations
    - Async web scraping with connection pooling
    - Real-time performance monitoring

  version: 1.0.0
  contact:
    name: Sectors Team
    email: human@supertype.ai
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://sectors-news-endpoint.fly.dev
    description: Production server
security:
  - ApiKeyAuth: []
...
paths:
...
  /articles:
    get:
      tags: [Articles]
      summary: Get news articles
      description: Retrieve news articles with optional filtering by subsector or ID
      security: []
      parameters:
        - name: subsector
          in: query
          description: Filter by subsector
          schema:
            type: string
          example: "banking"
        - name: sub_sector
          in: query
          description: Alternative parameter name for subsector
          schema:
            type: string
        - name: id
          in: query
          description: Filter by specific article ID
          schema:
            type: integer
          example: 123
      responses:
        "200":
          description: Successfully retrieved articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"

    post:
      tags: [Articles]
      summary: Insert news article
      description: Insert a single news article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsArticleInput"
      responses:
        "200":
          description: Article inserted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsertResponse"
        "400":
          description: Invalid input or duplicate source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags: [Articles]
      summary: Update news article
      description: Update an existing news article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/NewsArticleInput"
                - type: object
                  required: [id]
                  properties:
                    id:
                      type: integer
                      description: Article ID to update
      responses:
        "200":
          description: Article updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateResponse"

    delete:
      tags: [Articles]
      summary: Delete news articles
      description: Delete multiple news articles by ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id_list]
              properties:
                id_list:
                  type: array
                  items:
                    type: integer
                  description: List of article IDs to delete
      responses:
        "200":
          description: Articles deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Deleted"
  /articles/list:
    post:
      tags: [Articles]
      summary: Insert multiple news articles
      description: Insert multiple news articles in batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/NewsArticleInput"
      responses:
        "200":
          description: Articles processed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InsertResponse"
  /url-article:
    post:
      tags: [Articles]
      summary: Generate article from URL (Optimized)
      description: |
        Extract and process a news article from a URL using advanced AI processing.

        **Performance Features:**
        - Intelligent caching for repeated URLs (~90% faster for cache hits)
        - Parallel processing of classification and scoring (~40-60% faster)
        - Optimized web scraping with connection pooling
        - Real-time performance monitoring

        **Processing Pipeline:**
        1. Web scraping with fallback strategies
        2. Content extraction and preprocessing  
        3. LLM-based title and summary generation
        4. Parallel classification (tags, tickers, sentiment, dimensions)
        5. Article scoring based on Indonesian market relevance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source, timestamp]
              properties:
                source:
                  type: string
                  format: uri
                  description: URL of the news article to process (required, must not be null or empty)
                  example: "https://www.example.com/news/article"
                  minLength: 1
                timestamp:
                  type: string
                  format: date-time
                  description: Timestamp of the article (required, must not be null or empty)
                  example: "2024-01-15 10:30:00"
                  minLength: 1
      responses:
        "200":
          description: Article processed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "400":
          description: Validation error - missing or invalid required fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: "error"
                message: "Source URL is required"
        "500":
          description: Processing failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /url-article/post:
    post:
      tags: [Articles]
      summary: Post processed article
      description: Insert a previously processed article without regeneration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsArticleInput"
      responses:
        "200":
          description: Article posted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsertResponse"
  /evaluate-article:
    post:
      tags: [Articles]
      summary: Get article score
      description: Evaluate and score a news article based on Indonesian market relevance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
                  description: Article content to score
                title:
                  type: string
                  description: Article title
                tickers:
                  type: array
                  items:
                    type: string
                  description: Associated tickers
      responses:
        "200":
          description: Article scored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  score:
                    type: string
                    description: Article score (0-100+)
                    example: "75"
  /stock-split:
    post:
      tags: [Articles]
      summary: Insert stock split news
      description: Generate and insert stock split announcement articles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required: [symbol, date, split_ratio, updated_on, applied_on]
                properties:
                  symbol:
                    type: string
                    description: Stock ticker symbol
                    example: "BBCA"
                  date:
                    type: string
                    format: date
                    description: Effective date of split
                    example: "2024-02-01"
                  split_ratio:
                    type: number
                    description: Split ratio
                    example: 2
                  updated_on:
                    type: string
                    format: date
                    description: Last update date
                  applied_on:
                    type: string
                    format: date
                    description: Application date
      responses:
        "200":
          description: Stock split articles generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InsertResponse"
  /dividend:
    post:
      tags: [Articles]
      summary: Insert dividend news
      description: Generate and insert dividend announcement articles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                required:
                  [symbol, dividend_amount, ex_date, updated_on, payment_date]
                properties:
                  symbol:
                    type: string
                    description: Stock ticker symbol
                    example: "BBCA"
                  dividend_amount:
                    type: number
                    description: Dividend amount in IDR
                    example: 150
                  ex_date:
                    type: string
                    format: date
                    description: Ex-dividend date
                    example: "2024-02-15"
                  updated_on:
                    type: string
                    format: date
                    description: Last update date
                  payment_date:
                    type: string
                    format: date
                    description: Payment date
      responses:
        "200":
          description: Dividend articles generated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InsertResponse"

  # Filings Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
    NewsArticle:
      type: object
      properties:
        id:
          type: integer
          description: Unique article ID
        title:
          type: string
          description: Article title
        body:
          type: string
          description: Article content
        source:
          type: string
          format: uri
          description: Source URL
        timestamp:
          type: string
          format: date-time
          description: Article timestamp
        sector:
          type: string
          description: Industry sector
          example: "Financial Services"
        sub_sector:
          type: array
          items:
            type: string
          description: Industry subsectors
          example: ["banking"]
        tags:
          type: array
          items:
            type: string
          description: Content tags
          example: ["earnings", "positive", "financial results"]
        tickers:
          type: array
          items:
            type: string
          description: Associated stock tickers
          example: ["BBCA.JK", "BMRI.JK"]
        dimension:
          type: object
          description: Article dimensions analysis
          properties:
            valuation:
              type: integer
              nullable: true
            future:
              type: integer
              nullable: true
            technical:
              type: integer
              nullable: true
            financials:
              type: integer
              nullable: true
            dividend:
              type: integer
              nullable: true
            management:
              type: integer
              nullable: true
            ownership:
              type: integer
              nullable: true
            sustainability:
              type: integer
              nullable: true
        score:
          type: integer
          description: Article relevance score (0-150)
          example: 85
    NewsArticleInput:
      type: object
      properties:
        title:
          type: string
          description: Article title
        body:
          type: string
          description: Article content
        source:
          type: string
          format: uri
          description: Source URL
        timestamp:
          type: string
          format: date-time
          description: Article timestamp
        sector:
          type: string
          description: Industry sector
        sub_sector:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Industry subsector(s)
        subsector:
          type: string
          description: Alternative parameter for sub_sector
        tags:
          type: array
          items:
            type: string
          description: Content tags
        tickers:
          type: array
          items:
            type: string
          description: Associated stock tickers
        dimension:
          type: object
          description: Article dimensions
        score:
          type: integer
          description: Article score
...
    InsertResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, failed, restricted]
          description: Operation status
        id:
          type: integer
          description: Inserted record ID
        status_code:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Status message
        id_duplicate:
          type: integer
          description: ID of duplicate record (if restricted)
        error:
          type: string
          description: Error message (if failed)
...